Rename TeamArgs to LookingMessageInfo
