+++ src/LookingMessageInfo.ts
